% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthengine.R
\name{aggregate_ee}
\alias{aggregate_ee}
\title{Aggregate an Earth Engine Image Collection by period}
\usage{
aggregate_ee(object, frequency = "month", reduce_by = NULL)
}
\arguments{
\item{object}{\code{object}: a data object produced by \code{\link[=collect_ee]{collect_ee()}}}

\item{frequency}{\code{str}, \code{optional}: either \code{"month"} or \code{"year"} are accepted}

\item{reduce_by}{\code{str}, \code{optional}: summary statistic or technique to perform on
aggregated data. If NULL (default), will calculate the mean per period}
}
\value{
an updated \code{\link[=collect_ee]{collect_ee()}} object that can be passed on to \code{\link[=map_ee]{map_ee()}}
or \code{\link[=download_ee]{download_ee()}}
}
\description{
Performs simple temporal aggregation of Earth Engine Image collections by
month or year. For example, it can calculate the maximum normalized
difference vegetation index (NDVI) of an area of interest for each year. Note
that processing times can increase substantially with an increased number of
images.
}
\examples{

img <- collect_ee(
  collection = "LANDSAT/LC09/C02/T1_L2",
  coords = c(149.769345, -30.335861, 149.949173, -30.206271),
  date = "2021-06-01",
  end_date = "2022-06-01"
)

preprocess_ee(img, spectral = "NDVI")
# aggregate_ee(img, reduce_by = "median")

}
