% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_bbox}
\alias{config_bbox}
\title{Specify bounding box limits as a vector}
\usage{
config_bbox(x, boundaries = NULL)
}
\arguments{
\item{x}{\code{harvester} must be an object of this class, which can be created
using \code{load_config()}}

\item{boundaries}{\code{num} a vector of 4 numerical values representing the left,
bottom, right and top coordinates.}
}
\value{
\verb{list; harvester} a list object with additional class of "harvester".
This list stores configuration options for the data harvester.
}
\description{
Define the bounding box for the data collected. Normally rectangular, since
it is simple/useful for plotting.
}
\details{
For EPSG:4326 the bounding box is defined by (long_left, lat_bottom,
long_right, lat_top).

If projected coordinate reference system is used, the bounding box is defined
by: (easting_left, northing_bottom, easting_right, northing_top)

If \code{NULL}, the bounding box will be set at a default min/max of point data
+/- 0.05 deg (180 arcsec). arcsec). Also note that if an invalid vector is
provided i.e. not containing 4 values, the default bounding box will apply.
}
