% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_dates}
\alias{config_dates}
\title{Define download date and times, and summary interval}
\usage{
config_dates(x, dates = NULL, interval = "year")
}
\arguments{
\item{x}{\code{harvester} must be an object of this class, which can be created
using \code{load_config()}}

\item{dates}{\code{num} a numeric value or vector of year values, e.g. 2019}

\item{interval}{\code{chr} define the interval for data summary. Use "year",
"month" or "day"}
}
\value{
A list object which stores configuration options for the data
harvester.
}
\description{
Define download date and times, and summary interval
}
