% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_dea}
\alias{config_dea}
\title{Configure DEA resource download settings}
\usage{
config_dea(x, layers = NULL, summaries = NULL)
}
\arguments{
\item{summaries}{}
}
\description{
Configure DEA resource download settings
}
