% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_paths}
\alias{config_paths}
\title{Configure paths to input and output folders and files}
\usage{
config_paths(x, input = NULL, output = NULL, in_file = NULL)
}
\arguments{
\item{x}{Data object of class \code{harvester}, created by \code{load_config()}. The
function will not accept any other data object that has not been processed
this way.}

\item{input}{Path to folder containing input files used to determine
coordinates and other features to download. If \code{NULL}, will not update any
existing value.}

\item{output}{Path to folder to save results. Images, data frames and other
geospatial data outputs are saved it this folder for easy access. If
\code{NULL}, will not update any existing value.}

\item{in_file}{Path to specific file used to determine coordinates and other
features to download. If \code{NULL}, will not update the existing value.}
}
\value{
A list object which stores configuration options for the data
harvester.
}
\description{
This function changes the values for 'inpath', 'outpath' and 'infname' in the
data harvester configuration.
}
