% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_silo}
\alias{config_silo}
\title{Configure SILO resource download settings}
\usage{
config_silo(x, layers = NULL, summaries = NULL)
}
\arguments{
\item{summaries}{}
}
\description{
Configure SILO resource download settings
}
