% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_slga}
\alias{config_slga}
\title{Configure SLGA resource download settings}
\usage{
config_slga(x, layers = NULL, summaries = "mean")
}
\arguments{
\item{summaries}{}
}
\description{
Configure SLGA resource download settings
}
