% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_xy_names}
\alias{config_xy_names}
\title{Specify the column names of x and y variables in data file}
\usage{
config_xy_names(x, xvar = NULL, yvar = NULL)
}
\arguments{
\item{x}{\code{harvester} must be an object of this class, which can be created
using \code{load_config()}.}

\item{xvar}{\code{chr} column name of x coordinate stored in input data file, will
replace value in \code{colname_lng}.}

\item{yvar}{\code{chr} column name of y coordinate stored in input data file, will
replace value in \code{colname_lat}.}
}
\value{
\verb{list; harvester} a list object with additional class of "harvester".
This list stores configuration options for the data harvester.
}
\description{
For the data harvester to work, it needs coordinate data. Different users
will call these x and y coordinates under different names in their data file
e.g. \code{"Lon"}, \code{"Long"}, \code{"x"}, \code{"Easting"} or \code{"East"}.
}
\details{
Changing  allows the data harvester to identify the x and y coordinates
appropriately with the user's help. The values are saved to the options
\code{colname_lng} and \code{colname_lat} in the data harvester configuration.
}
\examples{
#' # load template config, then select references to coordinate columns
load_config() \%>\%
  config_xy_names("Long", "Lat")
}
