% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dd_source_python}
\alias{dd_source_python}
\title{Source python module stored in Data-Harvester Python folder}
\usage{
dd_source_python(modulename, rpackage, report = FALSE)
}
\arguments{
\item{report}{}
}
\description{
Source python module stored in Data-Harvester Python folder
}
