% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{download_dem}
\alias{download_dem}
\title{Download Geoscience Australia DEM grid data}
\usage{
download_dem(layer, out_path, bounding_box, resolution = 1)
}
\arguments{
\item{layer}{Layer name(s) to download as a vector of strings}

\item{out_path}{Path to an output/download folder, e.g. \code{'downloads/'}}

\item{bounding_box}{A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, \code{c(min_x, min_y, max_x, max_y)}}

\item{resolution}{An integer value for resolution, in arc seconds}
}
\value{
a list of filenames (after files have been downloaded or processed)
}
\description{
Wrapper funtion to get the layers from the Geoscience Australia DEM 1 arc
second grid.
}
