% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthengine.R
\name{download_ee}
\alias{download_ee}
\title{Download an Earth Engine Image or Image Collection}
\usage{
download_ee(
  object,
  bands = NULL,
  scale = NULL,
  outpath = NULL,
  out_format = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{object}{(object) a data object produced by \code{collect_ee()}}

\item{overwrite}{}
}
\description{
Images are saved as GeoTIFF (.tif) files containing geospatial data, unless
otherwise specified in \code{out_format}.
}
