% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{download_landscape}
\alias{download_landscape}
\title{Download from SLGA Landscape Attributes}
\usage{
download_landscape(layer, bounding_box, out_path, resolution = 3)
}
\arguments{
\item{layer}{Layer name(s) to download as a vector of strings}

\item{bounding_box}{A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, \code{c(min_x, min_y, max_x, max_y)}}

\item{out_path}{Path to an output/download folder, e.g. \code{'downloads/'}}

\item{resolution}{\verb{ r params(resolution)}}
}
\value{
a list of filenames (after files have been downloaded or processed)
}
\description{
Download from SLGA Landscape Attributes
}
