% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiometric.R
\name{download_radiometric}
\alias{download_radiometric}
\title{Download from Geosciences Australia's Radiometric Map of Australia}
\usage{
download_radiometric(
  layer,
  bounding_box,
  out_path,
  resolution = 1,
  crs = "EPSG:4326",
  format_out = "GeoTIFF"
)
}
\arguments{
\item{layer}{Layer name(s) to download as a vector of strings}

\item{bounding_box}{A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, \code{c(min_x, min_y, max_x, max_y)}}

\item{out_path}{Path to an output/download folder, e.g. \code{'downloads/'}}

\item{resolution}{An integer value for resolution, in arc seconds}

\item{crs}{Coordinate referencing system (CRS) to use. Currently, only accepts 'EPSG:4326'}

\item{format_out}{File format of downloaded file. Can only be \code{"GeoTIFF"} or
\code{"NetCDF"}. Defaults to \code{"GeoTIFF"}}
}
\value{
a list of filenames (after files have been downloaded or processed)
}
\description{
Download from Geosciences Australia's Radiometric Map of Australia
}
\details{
\subsection{Layers}{
\itemize{
\item \code{radmap2019_grid_dose_terr_awags_rad_2019}
\item \code{radmap2019_grid_dose_terr_filtered_awags_rad_2019}
\item \code{radmap2019_grid_k_conc_awags_rad_2019}
\item \code{radmap2019_grid_k_conc_filtered_awags_rad_2019}
\item \code{radmap2019_grid_th_conc_awags_rad_2019}
\item \code{radmap2019_grid_th_conc_filtered_awags_rad_2019}
\item \code{radmap2019_grid_thk_ratio_awags_rad_2019}
\item \code{radmap2019_grid_u2th_ratio_awags_rad_2019}
\item \code{radmap2019_grid_u_conc_awags_rad_2019}
\item \code{radmap2019_grid_u_conc_filtered_awags_rad_2019}
\item \code{radmap2019_grid_uk_ratio_awags_rad_2019}
\item \code{radmap2019_grid_uth_ratio_awags_rad_2019}
}
}
}
