% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slga.R
\name{download_slga}
\alias{download_slga}
\title{Download layers from SLGA data server}
\usage{
download_slga(
  layer,
  bounding_box,
  out_path,
  resolution = 3,
  depth_min = 0,
  depth_max = 200,
  get_ci = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{layer}{\code{string}: layer name(s) to download}

\item{bounding_box}{\code{numeric}: coordinates in EPSG:4326 use to generate a
bounding box around are of interest, i.e. \code{c(min_x, min_y, max_x, max_y)}}

\item{out_path}{\code{string}: path to output folder e.g. 'downloads/'}

\item{resolution}{\code{numeric}, \code{optional}: resolution in arcsec. Defaults to 3
arcsec (~ 90 m), which is the native resolution of SLGA data}

\item{depth_min}{\code{integer}, \code{optional}: SLGA layers are aggregated by depth.
Use \code{depth_min} and \code{depth_max} to define specific depths to collect}

\item{depth_max}{\code{integer}, \code{optional}: SLGA layers are aggregated by depth.
Use \code{depth_min} and \code{depth_max} to define specific depths to collect}

\item{get_ci}{\code{logical}, \code{optional}: download upper and lower 95\% confidence
limits with the the layer. Defaults to TRUE}

\item{verbose}{\code{logical}, \code{optional}: print as much information as possible
to the debug log. Defaults to FALSE}
}
\value{
path name(s) of layer(s) downloaded
}
\description{
Download layers from SLGA data server
}
\examples{
NULL
}
