% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthengine.R
\name{ee_download}
\alias{ee_download}
\title{Download an Earth Engine Image or Image Collection}
\usage{
ee_download(
  object,
  bands = NULL,
  scale = NULL,
  outpath = NULL,
  out_format = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{object}{(object) a data object produced by \code{ee_collect()}}

\item{overwrite}{}
}
\description{
Images are saved as GeoTIFF (.tif) files containing geospatial data, unless
otherwise specified in \code{out_format}.
}
