% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthengine.R
\name{ee_map}
\alias{ee_map}
\title{Visualise an Earth Engine Image or ImageCollection on an interactive map}
\usage{
ee_map(object, bands, minmax = NULL, palette = NULL)
}
\arguments{
\item{save_to}{}
}
\description{
Visualise an Earth Engine Image or ImageCollection on an interactive map
}
