% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_values}
\alias{extract_values}
\title{Extract values from GeoTIFF image(s)}
\usage{
extract_values(path, xy_coords, method = "simple")
}
\arguments{
\item{path}{path to a folder containing one or more GeoTIFF files e.g.
"~/Documents/rasters", or path pointing to one or more GeoTIFF files e.g.
c("~/rasters/clay.tif", "~/rasters/density.tif")}

\item{xy_coords}{a data frame of (x, y) or (longitude, latitude) coordinates.
Note the order - incorrect order of coordinates may result in erroneous or
NaN results.}

\item{method}{one of "simple", which returns the exact value at the cell or
pixel that the point falls in, or "bilinear", which returns a value
interpolated from the values of the four nearest raster cells.}
}
\value{
a data frame of values
}
\description{
Extract values from one or more GeoTIFF images based on a set of point
locations, supplied as (x, y) or (longitude, latitude) coordinates. If a path
to a folder is provided, all images in the folder will be processed.
}
