% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest}
\alias{harvest}
\title{A headless version of the Data-Harvester (with some limitations)}
\usage{
harvest(
  path_to_config,
  log_name = "download_log",
  preview = FALSE,
  contour = FALSE
)
}
\arguments{
\item{path_to_config}{\code{string}: path to YAML config file, e.g. "settings/"}

\item{log_name}{\code{string}: name of output log file which contains some
information about downloaded files}

\item{preview}{\code{logical}: preview rasters by plotting the first band only}

\item{contour}{\code{logical}: add contour lines to the plot. Defaults to FALSE}
}
\description{
The \code{\link[=harvest]{harvest()}} function requires a YAML configuration file to work and will
perform data aggregation, downloading and logging based on what is parsed
from the YAML file. For more information on 'headless' functionality, see the
documentation online (site not live yet). A YAML template can be generated
using the function \code{\link[=create_yaml]{create_yaml()}}
}
