% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ee.R
\name{harvest_ee}
\alias{harvest_ee}
\title{Start harvest instance}
\usage{
harvest_ee(
  config = NULL,
  collection = NULL,
  coords = NULL,
  date = NULL,
  end_date = NULL,
  buffer = NULL,
  bound = FALSE
)
}
\arguments{
\item{bound}{}
}
\description{
Start harvest instance
}
