% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_logtable}
\alias{init_logtable}
\title{Create a dataframe to store raster download and processing information}
\usage{
init_logtable()
}
\description{
This functions does not have any arguments. The dataframe can be updated with
the \code{\link[=update_logtable]{update_logtable()}} function.
}
