% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_start.R
\name{initialise}
\alias{initialise}
\title{Initialise and validate Data-Harvester and dependencies}
\usage{
initialise(env = NULL, earthengine = FALSE)
}
\arguments{
\item{env}{\code{chr} name of environment. If the environment doesn't currently
exist, one will be created in the same name using the default conda binary}

\item{earthengine}{\code{logical} initialise Earth Engine if TRUE. Defaults to FALSE}
}
\description{
Initialise and validate Data-Harvester and dependencies
}
