% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Llara}
\alias{Llara}
\title{Example data Llara}
\format{
A dataset containing x and y coordinates.
\describe{
\item{Lat}{latitude}
\item{Long}{longitude}
\item{Easting}{easting}
\item{Northing}{northing}
\item{Depth}{depth as range, in cm}
}
}
\usage{
Llara
}
\description{
A dataset.
}
\keyword{datasets}
