% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load a configuration file for data harvester}
\usage{
load_config(x = NULL, choose = 1)
}
\arguments{
\item{x}{Name of a config file (no .extenson), or path to ".yaml" config
file. If NULL, loads a default config file normally used to demonstrate the
use of the data harvester. This default configuration can also be used as a
template.}

\item{choose}{If more than one file with the same name is found, this
argument can be used to select the desired file by position. Defaults to
the first file listed.}
}
\value{
A list object which stores configuration options for the data
harvester.
}
\description{
This function performs a sequence of checks to read a \code{.yaml} configuration
file for the data harvester.
}
\details{
First, if no file name or file path is provided, it will load a template
configuration.

If only a config name is provided (without paths or extension), it will
search folders relative to the working folder for a file with the same name,
including common folders such as "data" and "config". Where multiple files
are found, the user can choose the right file to load by specifying it using
the argument \code{choose}.

If a file path is provided, it will load the file from the path.
}
\examples{
load_config() # loads template config as a start
}
