% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{load_settings}
\alias{load_settings}
\title{Load settings from YAML file}
\usage{
load_settings(path_to_yaml)
}
\arguments{
\item{path}{Path to settings file in YAML format/extension.}
}
\value{
A settings namespace object
}
\description{
Load settings from YAML file
}
