% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_rasters}
\alias{plot_rasters}
\title{Preview all images in a folder, recursively}
\usage{
plot_rasters(path, contour = FALSE, points = FALSE, x = NULL, y = NULL)
}
\arguments{
\item{path}{path to folder}

\item{contour}{show contour lines. Defaults to FALSE}
}
\description{
Preview all images in a folder, recursively
}
