% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ee.R
\name{preprocess_ee}
\alias{preprocess_ee}
\title{Preproces image collection}
\usage{
preprocess_ee(
  ee,
  mask_clouds = TRUE,
  reduce = "median",
  spectral = NULL,
  clip = TRUE
)
}
\arguments{
\item{spectral}{}
}
\description{
Preproces image collection
}
