% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.harvester}
\alias{print.harvester}
\title{Print harvester objects}
\usage{
\method{print}{harvester}(x, ...)
}
\arguments{
\item{x}{\code{harvester} must be an object of this class, which can be created
using \code{load_config()}}

\item{...}{other arguments passed to function}
}
\value{
the same input object
}
\description{
Print harvester objects
}
