% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rstudio_viewer}
\alias{rstudio_viewer}
\title{View an HTML file in the RStudio Viewer pane.}
\usage{
rstudio_viewer(file_name, file_path = NULL)
}
\arguments{
\item{file_name}{The HTML file to be viewed.}

\item{file_path}{The path to the directory in which the HTML file can be
found, if the HTML file to be viewed is not in the current working
directory.}
}
\description{
View an HTML file (e.g., a rendered R Markdown document) in the RStudio
Viewer pane without re-rendering the R Markdown document.
}
