% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{validate_conda}
\alias{validate_conda}
\title{Check if conda is available}
\usage{
validate_conda(reinstall = FALSE)
}
\arguments{
\item{reinstall}{Force re-install of miniconda. Defaults to FALSE}
}
\description{
Check if conda is available
}
