% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{validate_env}
\alias{validate_env}
\title{Load default conda environment}
\usage{
validate_env(env = NULL)
}
\arguments{
\item{env}{\code{chr} name of conda environment to load. Defaults to \code{NULL}, which
automatically searches for the environments \code{geopy} and \code{dataharvestR}}
}
\description{
Load default conda environment
}
