% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{validate_py_packages}
\alias{validate_py_packages}
\title{Check if required Python packages for Data-Harvester exist.}
\usage{
validate_py_packages(env = NULL)
}
\arguments{
\item{env}{\code{chr} name of environment}
}
\value{
\code{logical}
}
\description{
Internal function. First, the function checks if all required packages have
been installed. Then it does a quick check to see if gdal is at version
3.4.2. If any package appears to be missing they will be reinstalled.
}
