% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{authenticate_ee}
\alias{authenticate_ee}
\title{Authenticate to Google Earth Engine API}
\usage{
authenticate_ee()
}
\description{
Utilises google-cloud-sdk to initialise and authenticate to the Earth Engine
API. An API token containing the user's credentials is saved locally and can
be used to authenticate vial Application Default Credentials.
}
