% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_crs}
\alias{config_crs}
\title{Define coordinate reference system}
\usage{
config_crs(x, target_crs = "EPSG:4326")
}
\arguments{
\item{target_crs}{}
}
\description{
Define coordinate reference system
}
