% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_dates}
\alias{config_dates}
\title{Define download date and times, and summary interval}
\usage{
config_dates(x, dates = NULL, interval = "year")
}
\arguments{
\item{interval}{}
}
\description{
Define download date and times, and summary interval
}
