% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_dem}
\alias{config_dem}
\title{Configure DEM resource download settings}
\usage{
config_dem(x, layers = NULL, summaries = NULL)
}
\arguments{
\item{summaries}{}
}
\description{
Configure DEM resource download settings
}
