% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_paths}
\alias{config_paths}
\title{Configure paths to input and output folders and files}
\usage{
config_paths(x, input = NULL, output = NULL, in_file = NULL)
}
\arguments{
\item{x}{\code{harvester} must be an object of this class, which can be created
using \code{load_config()}.}

\item{input}{\code{chr} path to folder containing input files used to determine
coordinates and other features to download. If \code{NULL}, does nothing.}

\item{output}{\code{chr} path to folder to save results. Images, data frames and
other geospatial data outputs are saved it this folder for easy access. If
\code{NULL}, does nothing.}

\item{in_file}{\code{chr} path to specific file used to determine coordinates and
other features to download. If \code{NULL}, does nothing.}
}
\value{
\verb{list; harvester} a list object with additional class of "harvester".
This list stores configuration options for the data harvester.
}
\description{
This function changes the values for \code{inpath}, \code{outpath} and \code{infname} in the
data harvester configuration.
}
