% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_res}
\alias{config_res}
\title{Define resolution in arcsecs}
\usage{
config_res(x, value = NULL)
}
\arguments{
\item{x}{\code{harvester} must be an object of this class, which can be created
using \code{load_config()}}

\item{value}{\code{num} resolution in arcsecs}
}
\value{
A list object which stores configuration options for the data
harvester.
}
\description{
Define resolution in arcsecs
}
