% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_silo}
\alias{config_silo}
\title{Configure SILO resource download settings}
\usage{
config_silo(x, layers = NULL, summaries = NULL)
}
\arguments{
\item{x}{\code{harvester} must be an object of this class, which can be created
using \code{load_config()}}

\item{layers}{\code{chr} name of layer. This is very specific to the SLGA list of
layers that can be downloaded.}

\item{summaries}{\code{chr} for each layer specified, a summary needs to be paired
to it. If only one summary value is provided, the function will assume that
it is for all layers}
}
\value{
A list object which stores configuration options for the data
harvester.
}
\description{
Configure SILO resource download settings
}
