% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dd_source_python}
\alias{dd_source_python}
\title{(Internal) Source module stored in python folder}
\usage{
dd_source_python(modulename, rpackage, report = FALSE)
}
\arguments{
\item{report}{}
}
\description{
(Internal) Source module stored in python folder
}
