% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{download_dea}
\alias{download_dea}
\title{Download from Digital Earth Australia}
\usage{
download_dea(
  layer,
  bounding_box,
  out_path,
  years,
  resolution,
  crs = "EPSG:4326",
  format_out = "GeoTIFF"
)
}
\arguments{
\item{layer}{Layer name(s) to download as a vector of strings}

\item{bounding_box}{A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, \code{c(min_x, min_y, max_x, max_y)}}

\item{out_path}{Path to an output/download folder, e.g. \code{'downloads/'}}

\item{years}{Year(s) to download data from (if available), e.g. \code{2019:2021} or \code{c(2015, 2020)}}

\item{resolution}{An integer value for resolution, in arc seconds}

\item{crs}{Coordinate referencing system (CRS) to use. Currently, only accepts 'EPSG:4326'}

\item{format_out}{Output format, either "GeoTIFF" or "NetCDF". Defaults to
"GeoTIFF"}
}
\value{
a list of filenames (after files have been downloaded or processed)
}
\description{
Wrapper funtion to get the layers from Digital Earth Australia collections.
}
