% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{download_dem}
\alias{download_dem}
\title{Download Geoscience Australia DEM grid data}
\usage{
download_dem(layer, out_path, bounding_box, zresolution = 1)
}
\arguments{
\item{layer}{\code{string}: list of layer names to download}

\item{out_path}{\code{string}:}

\item{bounding_box}{\code{numeric}:}

\item{resolution}{\code{integer}, \code{optional}:}
}
\value{
a list of output filenames
}
\description{
Wrapper funtion to get the layers from the Geoscience Australia DEM 1 arc
second grid.
}
