% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{download_landscape}
\alias{download_landscape}
\title{Download from SLGA Landscape Attributes}
\usage{
download_landscape(layer, bounding_box, out_path, resolution = 3)
}
\arguments{
\item{layer}{Layer name(s) to download as a vector of strings}

\item{bounding_box}{A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, \code{c(min_x, min_y, max_x, max_y)}}

\item{out_path}{Path to an output/download folder, e.g. \code{'downloads/'}}

\item{resolution}{\verb{ r params(resolution)}}
}
\value{
a list of filenames (after files have been downloaded or processed)
}
\description{
Download from SLGA Landscape Attributes
}
\details{
\subsection{Layers}{
\itemize{
\item \code{Prescott_index}
\item \code{net_radiation_jan}
\item \code{net_radiation_july}
\item \code{total_shortwave_sloping_surf_jan}
\item \code{total_shortwave_sloping_surf_july}
\item \code{Slope}
\item \code{Slope_median_300m}
\item \code{Slope_relief_class}
\item \code{Aspect}
\item \code{Relief_1000m}
\item \code{Relief_300m}
\item \code{Topographic_wetness_index}
\item \code{TPI_mask}
\item \code{SRTM_TopographicPositionIndex}
\item \code{Contributing_area}
\item \code{MrVBF}
\item \code{Plan_curvature}
\item \code{Profile_curvature}
}
}
}
