% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo.R
\name{download_silo}
\alias{download_silo}
\title{Download Australian climate data from SILO}
\usage{
download_silo(
  layername,
  years,
  out_path,
  bounding_box = NULL,
  format_out = "nc",
  delete_temp = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Download Australian climate data from SILO
}
