% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo.R
\name{download_silo}
\alias{download_silo}
\title{Download from SILO database}
\usage{
download_silo(
  layer,
  bounding_box,
  out_path,
  years,
  format_out = "tif",
  delete_temp = FALSE
)
}
\arguments{
\item{layer}{Layer name(s) to download as a vector of strings}

\item{bounding_box}{A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, \code{c(min_x, min_y, max_x, max_y)}}

\item{out_path}{Path to an output/download folder, e.g. \code{'downloads/'}}

\item{years}{Year(s) to download data from (if available), e.g. \code{2019:2021} or \code{c(2015, 2020)}}

\item{format_out}{}

\item{delete_temp}{Delete any temporary files that were created. Defaults to \code{FALSE}}
}
\value{
a list of filenames (after files have been downloaded or processed)
}
\description{
Download from SILO database
}
