% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slga.R
\name{download_slga}
\alias{download_slga}
\title{Download layers from SLGA data server}
\usage{
download_slga(
  layer,
  bounding_box,
  out_path,
  resolution = 3,
  depth_min = 0,
  depth_max = 200,
  get_ci = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Download layers from SLGA data server
}
