% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slga.R
\name{download_slga}
\alias{download_slga}
\title{Download from SLGA (Soil Atributes)}
\usage{
download_slga(
  layer,
  bounding_box,
  out_path,
  resolution = 3,
  depth_min = 0,
  depth_max = 200,
  get_ci = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{layer}{Layer name(s) to download as a vector of strings}

\item{bounding_box}{A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, \code{c(min_x, min_y, max_x, max_y)}}

\item{out_path}{Path to an output/download folder, e.g. \code{'downloads/'}}

\item{resolution}{An integer value for resolution, in arc seconds}

\item{depth_min, depth_max}{SLGA layers can be filtered between specific
depths. Minimum and maximum depths can be set using \code{depth_min} and
\code{depth_max}. Defaults to \code{0} and \code{200}, respectively}

\item{get_ci}{Also download upper and lower 95\% confidence limits with the
the layer. Defaults to \code{TRUE}}

\item{verbose}{Print as much information as possible to the debug log}
}
\value{
a list of filenames (after files have been downloaded or processed)
}
\description{
Wrapper funtion to get layers from the Soil and Landscape Grid of Australia
(SLGA).
}
\details{
\subsection{Layers}{
\itemize{
\item \code{Bulk_Density}
\item \code{Organic_Carbon}
\item \code{Clay}
\item \code{Silt}
\item \code{Sand}
\item \code{pH_CaCl2}
\item \code{Available_Water_Capacity}
\item \code{Total_Nitrogen}
\item \code{Total_Phosphorus}
\item \code{Effective_Cation_Exchange_Capacity}
\item \code{Depth_of_Regolith}
\item \code{Depth_of_Soil}
}
}
}
