% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthengine.R
\name{ee_preprocess}
\alias{ee_preprocess}
\title{Preprocess an Earth Engine Image or Image Collection}
\usage{
ee_preprocess(
  object,
  mask_clouds = TRUE,
  reduce = "median",
  spectral = NULL,
  clip = TRUE
)
}
\arguments{
\item{object}{(object) a data object produced by \code{ee_collect()}}

\item{mask_clouds}{(logical, optional) if TRUE, will perform cloud and shadow
masking on image(s). Defaults to TRUE}

\item{reduce}{(str, optional) Summary technique used to reduce an image
collection to a single composite. Defaults to "median"}

\item{spectral}{(logical, optional) If TRUE, will automatically calculate
spectral index based on \href{https://github.com/awesome-spectral-indices/awesome-spectral-indices}{Awesome Spectral Indices}.
If required bands are not available, the calculation will be skipped.
Defaults to NULL}

\item{clip}{(logical, optional)}
}
\value{
an updated \code{ee_collect()} object that can be passed on to \code{ee_map()},
\code{ee_aggregate()} or \code{ee_download()}
}
\description{
Obtain image stacks from a Google Earth Engine catalog collection for
processing. Full support for Sentinel-2, Sentinel-3, Landsat 5-9 and most
MODIS satellites. Preprocessing performs server-side filtering, cloud
masking, scaling and offsetting, calculation of spectral indices and
compositing into a single image representing, for example, the median, min,
max, mean, quantile or standard deviation of the images. For unsupported
collections, certain functions like scaling, offsetting, spectral indices and
cloud/shadow masking may not be available. Must be used on an object created
by the function, \code{ee_collect()}
}
\examples{

img <- ee_collect(collection = "LANDSAT/LC09/C02/T1_L2",
                  coords = c(149.769345,-30.335861, 149.949173,-30.206271),
                  date = "2021-06-01",
                  end_date = "2022-06-01")

ee_preprocess(img, spectral = "NDVI")

}
