% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_start.R
\name{harvest_start}
\alias{harvest_start}
\title{Initialise and validate Data-Harvester and dependencies}
\usage{
harvest_start(env = NULL)
}
\arguments{
\item{env}{\code{chr} name of environment. If the environment doesn't currently
exist, one will be created in the same name using the default conda binary.}
}
\description{
Initialise and validate Data-Harvester and dependencies
}
