% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise_harvester}
\alias{initialise_harvester}
\title{Initialise and validate Data-Harvester, including dependencies}
\usage{
initialise_harvester(earthengine = TRUE)
}
\arguments{
\item{earthengine}{\code{logical} initialise Earth Engine if TRUE. Defaults to FALSE}
}
\description{
Initialise and validate Data-Harvester, including dependencies
}
