% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise_harvester}
\alias{initialise_harvester}
\title{Initialise and validate Data-Harvester, including dependencies}
\usage{
initialise_harvester(envname = NULL, earthengine = FALSE, auth_mode = "gcloud")
}
\arguments{
\item{envname}{Use this Conda environment. Defaults to \code{"r-reticulate"}}

\item{earthengine}{Initialise Earth Engine if \code{TRUE.} Defaults to \code{FALSE}}

\item{auth_mode}{Authentication mode to access the GEE API. Using \code{"gcloud"}
normally works. The remaining three options are identical (\verb{"notebook", "rstudiocloud", "binder"}), just named differently for context}
}
\description{
Initialise and validate Data-Harvester, including dependencies
}
