% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthengine.R
\name{map_ee}
\alias{map_ee}
\title{Visualise an Earth Engine Image or Image Collection on an interactive map}
\usage{
map_ee(object, bands, minmax = NULL, palette = NULL)
}
\arguments{
\item{object}{\code{object}: a data object produced by \code{\link[=collect_ee]{collect_ee()}}}

\item{palette}{}
}
\description{
A \href{http://python-visualization.github.io/folium/}{folium} map is produced and
image(s) collected so far are displayed as layer(s) on top of the map.
}
