% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthengine.R
\name{preprocess_ee}
\alias{preprocess_ee}
\title{Preprocess an Earth Engine Image or Image Collection}
\usage{
preprocess_ee(
  object,
  mask_clouds = TRUE,
  reduce = "median",
  spectral = NULL,
  clip = TRUE
)
}
\arguments{
\item{object}{\code{object}: a data object produced by \code{\link[=collect_ee]{collect_ee()}}}

\item{mask_clouds}{\code{logical}, \code{optional}: perform cloud and shadow masking on
image(s). Defaults to TRUE}

\item{reduce}{\code{string}, \code{optional}: summary technique used to reduce an image
collection to a single composite. Defaults to "median"}

\item{spectral}{\code{logical}, \code{optional}: automatically calculate spectral index
based on \href{https://is.gd/T1ogFV}{Awesome Spectral Indices}. If required
bands are not available, the calculation will be skipped. Defaults to NULL}

\item{clip}{\code{logical}, \code{optional}: clip the image (removing surrounding
data). This mostly affects visualisation - the final download will always
respect the bounding box set by \code{coords}, regardless of this value.
Defaults to TRUE}
}
\value{
an updated \code{\link[=collect_ee]{collect_ee()}} object that can be passed on to \code{\link[=map_ee]{map_ee()}},
\code{\link[=aggregate_ee]{aggregate_ee()}} or \code{\link[=download_ee]{download_ee()}}
}
\description{
Obtain image stacks from a Google Earth Engine catalog collection for
processing. Full support for Sentinel-2, Sentinel-3, Landsat 5-9 and most
MODIS satellites. Preprocessing performs server-side filtering, cloud
masking, scaling and offsetting, calculation of spectral indices and
compositing into a single image representing, for example, the median, min,
max, mean, quantile or standard deviation of the images. For unsupported
collections, certain functions like scaling, offsetting, spectral indices and
cloud/shadow masking may not be available. Must be used on an object created
by the function, \code{\link[=collect_ee]{collect_ee()}}
}
\examples{

img <- collect_ee(
  collection = "LANDSAT/LC09/C02/T1_L2",
  coords = c(149.769345, -30.335861, 149.949173, -30.206271),
  date = "2021-06-01",
  end_date = "2022-06-01"
)

preprocess_ee(img, spectral = "NDVI")

}
