% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_logtable}
\alias{update_logtable}
\title{Update download and processing dataframe}
\usage{
update_logtable(
  logname,
  file_name,
  layer,
  source,
  settings,
  layertitles = list(),
  agfunctions = list(),
  loginfos = list()
)
}
\arguments{
\item{loginfos}{}
}
\description{
Update the dataframe table with the information from the raster download or
processing. The dataframe is simultaneoulsy saved to a csv file in default
output directory.
}
