% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{validate_conda}
\alias{validate_conda}
\title{Check if conda is available}
\usage{
validate_conda(reply = interactive())
}
\arguments{
\item{reply}{\code{logical} if \code{interactive()}, function will prompt user for
response. Defaults to TRUE when \code{interactive()} is chosen.}
}
\description{
Check if conda is available
}
