% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{wrap_config}
\alias{wrap_config}
\title{Wrap configuration file to data frame}
\usage{
wrap_config(df, config)
}
\arguments{
\item{df}{\code{data.frame} source data, containing x and y variables, usually as
a data frame object}

\item{config}{\code{chr} name of config file}
}
\value{
\verb{list; harvester} a list object with additional class of
"agrefed.wrap". This list stores both the configuration options and the
source datafor the data harvester.
}
\description{
Wrap configuration file to data frame
}
